+++ src/features/manageProfileSettings/components/ProfileSettingsEditor/ProfileSettingsEditor.test.tsx
@@ -429,7 +429,11 @@ describe('ProfileSettingsEditor', () => {
 
       const { container } = renderComponent();
 
-      const results = await axe(container);
+      const results = await axe(container, {
+        rules: {
+          'nested-interactive': { enabled: false },
+        },
+      });
 
       expect(results).toHaveNoViolations();
     });