+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Fix field collision in the record generation. Fixes [UILD-838].
 * Update language used for importing works/instances. Refs [UILD-839].
 * Fix multiple settings per profile. Refs [UILD-780].
+* Add accessibility tests. Refs [UILD-841].
 
 [UILD-744]:https://folio-org.atlassian.net/browse/UILD-744
 [UILD-816]:https://folio-org.atlassian.net/browse/UILD-816
@@ -23,6 +24,7 @@
 [UILD-838]:https://folio-org.atlassian.net/browse/UILD-838
 [UILD-839]:https://folio-org.atlassian.net/browse/UILD-839
 [UILD-780]:https://folio-org.atlassian.net/browse/UILD-780
+[UILD-841]:https://folio-org.atlassian.net/browse/UILD-841
 
 ## 2.0.4 (2026-06-03)
 * Fix default profile type persistence across edit form and profile settings. Fixes [UILD-820].